requirements.txt
asgiref==3.7.2
contourpy==1.2.1
cycler==0.12.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-interface==0.26.0
django-colorfield==0.11.0
django-import-export==3.2.0
djangoql==0.17.1
et-xmlfile==1.1.0
fonttools==4.51.0
kiwisolver==1.4.5
matplotlib==3.8.4
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
ply==3.11
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.6.1
text-unidecode==1.3
typing_extensions==4.7.1
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
